backend/requirements.txt
fastapi
uvicorn
supabase
requests
python-dotenv
packaging
PyJWT
bcrypt
fpdf2
python-multipart
